requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiokafka==0.12.0
aiosignal==1.3.2
alembic==1.8.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==24.3.0
autots==0.6.17
babel==2.17.0
bcrypt==4.0.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
boto3==1.35.98
botocore==1.35.98
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.3
cloudpickle==2.2.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
databricks-sdk==0.40.0
debugpy==1.8.13
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
dynaconf==3.2.10
et_xmlfile==2.0.0
evidently==0.6.7
executing==2.1.0
Faker==37.1.0
fastapi==0.109.2
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.0
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
gevent==24.11.1
geventhttpclient==2.0.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
googleapis-common-protos==1.66.0
GPUtil==1.4.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
h2o==3.46.0.6
h5py==3.13.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-resources==5.13.0
importlib_metadata==8.5.0
ipinfo==5.1.1
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lightgbm==4.5.0
litestar==2.15.2
litestar-htmx==0.4.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mat73==0.65
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mlflow==2.19.0
mlflow-skinny==2.19.0
mlserver==1.4.0
mlserver-mlflow==1.6.1
msgspec==0.19.0
multidict==6.1.0
multipart==1.2.1
mypy-extensions==1.0.0
narwhals==1.20.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
nvidia-nccl-cu12==2.23.4
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-grpc==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
optuna==4.1.0
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
plotly==5.24.1
pmdarima==2.0.4
polyfactory==2.20.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-grpc-prometheus==0.8.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.7.0
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.7.0
PyMySQL==1.1.1
pyparsing==3.2.0
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-rapidjson==1.14
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-click==1.8.8
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
scikit-learn==1.5.2
scipy==1.14.1
secure==0.3.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
smbprotocol==1.15.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.36.3
starlette_exporter==0.23.0
statsmodels==0.14.4
streamlit==1.41.1
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tritonclient==2.42.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.3.0
uuid6==2024.7.10
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.0
xgboost==2.1.3
yarl==1.18.3
zenml==0.71.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
